+++ pkg/plugin/plugin.go
@@ -74,8 +74,8 @@ type Metadata struct {
 	// automatic setting of HELM_HOST.
 	UseTunnel bool `json:"useTunnel"`
 
-	// PlatformHooks are commands that will run on events based on the platform specified.
-	PlatformHooks map[string]Hooks `json:"hooks"`
+	// PlatformTasks are commands that will run on events based on the platform specified.
+	PlatformTasks map[string]Tasks `json:"tasks"`
 
 	// Downloaders field is used if the plugin supply downloader mechanism
 	// for special protocols.